Fix getting IDs for existing authors
Once an author is registered, it will receive an ID. That ID must be used
in all session. That was no the case, but the ID would depend on the author's
order of joining in each session.

This leads to author ID not matching author data

[5.x]

ERM40414

Change-Id: Iee1df778c50c479f8f11cff2f10b35fe4b7a0ad3
(cherry picked from commit 03a38fe55611eab8c562281db2b6ee932d5aa514)